pyproject.toml
[project]
name = "exceltodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "psqlpy>=0.11.11",
    "python-dotenv>=1.2.2",
]
